refactor: scene asset 로딩 유틸리티를 lib/utils/scene_asset_loader.dart로 분리
## 변경
- lib/utils/scene_asset_loader.dart 신규 (206줄)
  - SceneAssetLoader 클래스로 매니페스트 로딩 + 디렉토리 매칭 + 정렬 캡슐화
  - _StoryHomeScreenState에 흩어져 있던 5개 private 메소드와
    4개 정적 RegExp, 2개 캐시 필드를 한 곳으로 통합
  - loadForEvent(event) / loadForTitle(title: ..., code: ...) public API
- lib/screens/story_home_screen.dart
  - scene 관련 필드/메소드 전부 제거 (약 178줄 감소)
  - 대신 final _sceneAssetLoader = SceneAssetLoader() 하나만 유지
  - _openEventDetailPage에서 _sceneAssetLoader.loadForEvent(event) 호출
- lib/data/user_repository.dart: 불필요한 dart:typed_data import 제거

## 결과
- story_home_screen.dart: 1,290줄 → 1,101줄 (-189줄)
- scene 로딩 로직이 테스트 가능한 독립 단위로 분리됨 (향후 단위 테스트 추가 가능)
