+++ lib/screens/story_home_screen.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:convert';
 import 'dart:math' as math;
 
 import 'package:flutter/material.dart';
@@ -14,6 +13,7 @@ import '../models/quiz_question.dart';
 import '../state/auth_providers.dart';
 import '../state/story_controller.dart';
 import '../state/story_state.dart';
+import '../utils/scene_asset_loader.dart';
 import '../widgets/bible_reader_page.dart';
 import '../widgets/event_detail_page.dart';
 import '../widgets/person_panel.dart';
@@ -39,25 +39,14 @@ class _StoryHomeScreenState extends ConsumerState<StoryHomeScreen> {
   final ScrollController _selectionPanelScrollController = ScrollController();
   final GlobalKey<ProfileTabPageState> _profileTabKey =
       GlobalKey<ProfileTabPageState>();
+  final SceneAssetLoader _sceneAssetLoader = SceneAssetLoader();
   ProviderSubscription<User?>? _authUserSubscription;
-  static final RegExp _sceneFilenamePattern = RegExp(
-    r'/scene_(\d+)\.(?:png|jpe?g|webp)$',
-    caseSensitive: false,
-  );
-  static final RegExp _sceneCodeDigitsPattern = RegExp(r'(\d+)$');
-  static final RegExp _sceneInvalidDirChars = RegExp(r'[\\/:*?"<>|]+');
-  static final RegExp _sceneWhitespacePattern = RegExp(r'\s+');
-  static final RegExp _sceneLooseNormalizePattern = RegExp(
-    r"[\s_\-:·,./\\(){}\[\]']+",
-  );
   PersonSortMode _personSortMode = PersonSortMode.eraOrder;
   int _selectionStep = 1;
   StorySelectionPanelStage _selectionPanelStage =
       StorySelectionPanelStage.expanded;
   double _selectionSheetExtent = _selectionSheetExpandedSize;
   Set<String> _draftSelectedPersonIds = <String>{};
-  List<String>? _assetManifestCache;
-  final Map<String, List<String>> _sceneAssetsCache = <String, List<String>>{};
 
   @override
   void initState() {
@@ -565,7 +554,7 @@ class _StoryHomeScreenState extends ConsumerState<StoryHomeScreen> {
   }
 
   Future<void> _openEventDetailPage(StoryEvent event) async {
-    final sceneAssetsFuture = _loadSceneAssetsForEvent(event);
+    final sceneAssetsFuture = _sceneAssetLoader.loadForEvent(event);
     if (!mounted) {
       return;
     }
@@ -608,184 +597,6 @@ class _StoryHomeScreenState extends ConsumerState<StoryHomeScreen> {
         ),
       ),
     );
-  }
-
-  Future<List<String>> _loadAssetManifest() async {
-    final cached = _assetManifestCache;
-    if (cached != null) {
-      return cached;
-    }
-
-    try {
-      final manifest = await AssetManifest.loadFromAssetBundle(rootBundle);
-      final assetKeys = manifest.listAssets();
-      _assetManifestCache = assetKeys;
-      return assetKeys;
-    } catch (_) {
-      // Fall through to JSON manifest for older/alternate environments.
-    }
-
-    try {
-      final rawManifest = await rootBundle.loadString('AssetManifest.json');
-      final decoded = json.decode(rawManifest);
-      if (decoded is Map<String, dynamic>) {
-        final assetKeys = decoded.keys.toList(growable: false);
-        _assetManifestCache = assetKeys;
-        return assetKeys;
-      }
-    } catch (_) {
-      // Return empty manifest when assets are unavailable in current build.
-    }
-    _assetManifestCache = const <String>[];
-    return _assetManifestCache!;
-  }
-
-  String _sceneDirectoryNameForTitle(String title) {
-    final replaced = title.replaceAll(_sceneInvalidDirChars, '_').trim();
-    final trimmedDots = replaced.replaceAll(RegExp(r'^\.+|\.+$'), '');
-    final collapsed = trimmedDots
-        .replaceAll(_sceneWhitespacePattern, ' ')
-        .trim();
-    return collapsed.isNotEmpty ? collapsed : 'untitled_event';
-  }
-
-  String _normalizeSceneLookupKey(String text) {
-    return text
-        .toLowerCase()
-        .replaceAll(_sceneLooseNormalizePattern, '')
-        .trim();
-  }
-
-  String _stripSceneDirectoryPrefix(String directoryName) {
-    return directoryName.replaceFirst(RegExp(r'^\d+\s*'), '').trim();
-  }
-
-  String? _scenePrefixForCode(String code) {
-    final match = _sceneCodeDigitsPattern.firstMatch(code.trim());
-    final digits = match?.group(1);
-    if (digits == null || digits.isEmpty) {
-      return null;
-    }
-    final numeric = int.tryParse(digits);
-    if (numeric == null) {
-      return null;
-    }
-    return numeric.toString().padLeft(3, '0');
-  }
-
-  Future<List<String>> _loadSceneAssetsForEvent(StoryEvent event) async {
-    return _loadSceneAssetsForLookup(title: event.title, code: event.code);
-  }
-
-  Future<List<String>> _loadSceneAssetsForLookup({
-    required String title,
-    String? code,
-  }) async {
-    final dirName = _sceneDirectoryNameForTitle(title);
-    final cached = _sceneAssetsCache[dirName];
-    if (cached != null) {
-      return cached;
-    }
-
-    final manifest = await _loadAssetManifest();
-    const sceneRoot = 'assets/story_images_thumbs/';
-    final allScenePaths = manifest
-        .where(
-          (path) =>
-              path.startsWith(sceneRoot) &&
-              _sceneFilenamePattern.hasMatch(path),
-        )
-        .toList(growable: false);
-
-    final knownDirs = <String>{};
-    for (final path in allScenePaths) {
-      final relative = path.substring(sceneRoot.length);
-      final slashIndex = relative.indexOf('/');
-      if (slashIndex <= 0) {
-        continue;
-      }
-      knownDirs.add(relative.substring(0, slashIndex));
-    }
-
-    var chosenDir = dirName;
-    final directPrefix = '$sceneRoot$chosenDir/';
-    final hasDirect = allScenePaths.any(
-      (path) => path.startsWith(directPrefix),
-    );
-    if (!hasDirect) {
-      final codePrefix = code == null ? null : _scenePrefixForCode(code);
-      if (codePrefix != null) {
-        final codeMatchedDir =
-            knownDirs
-                .where((dir) => dir.startsWith('$codePrefix '))
-                .toList(growable: false)
-              ..sort((a, b) => a.length.compareTo(b.length));
-        if (codeMatchedDir.isNotEmpty) {
-          chosenDir = codeMatchedDir.first;
-        }
-      }
-    }
-
-    final chosenPrefixAfterCode = '$sceneRoot$chosenDir/';
-    final hasCodeMatched = allScenePaths.any(
-      (path) => path.startsWith(chosenPrefixAfterCode),
-    );
-    if (!hasCodeMatched) {
-      final titleKey = _normalizeSceneLookupKey(title);
-      final dirNameKey = _normalizeSceneLookupKey(dirName);
-      final fallbackCandidates =
-          knownDirs
-              .map((dir) {
-                final rawKey = _normalizeSceneLookupKey(dir);
-                final strippedKey = _normalizeSceneLookupKey(
-                  _stripSceneDirectoryPrefix(dir),
-                );
-                var score = -1;
-                if (rawKey == titleKey || rawKey == dirNameKey) {
-                  score = 0;
-                } else if (strippedKey == titleKey ||
-                    strippedKey == dirNameKey) {
-                  score = 1;
-                } else if (strippedKey.endsWith(titleKey) ||
-                    strippedKey.endsWith(dirNameKey)) {
-                  score = 2;
-                } else if (strippedKey.contains(titleKey) ||
-                    strippedKey.contains(dirNameKey) ||
-                    titleKey.contains(strippedKey) ||
-                    dirNameKey.contains(strippedKey)) {
-                  score = 3;
-                }
-                return (dir: dir, score: score);
-              })
-              .where((candidate) => candidate.score >= 0)
-              .toList(growable: false)
-            ..sort((a, b) {
-              final byScore = a.score.compareTo(b.score);
-              if (byScore != 0) {
-                return byScore;
-              }
-              return a.dir.length.compareTo(b.dir.length);
-            });
-      if (fallbackCandidates.isNotEmpty) {
-        chosenDir = fallbackCandidates.first.dir;
-      }
-    }
-
-    final chosenPrefix = '$sceneRoot$chosenDir/';
-    final sceneAssets =
-        allScenePaths
-            .where((path) => path.startsWith(chosenPrefix))
-            .toList(growable: false)
-          ..sort((a, b) {
-            final aMatch = _sceneFilenamePattern.firstMatch(a);
-            final bMatch = _sceneFilenamePattern.firstMatch(b);
-            final aIndex = int.tryParse(aMatch?.group(1) ?? '') ?? 0;
-            final bIndex = int.tryParse(bMatch?.group(1) ?? '') ?? 0;
-            return aIndex.compareTo(bIndex);
-          });
-
-    _sceneAssetsCache[dirName] = sceneAssets;
-    return sceneAssets;
   }
 
   String _eraTestament(Era era) {
